feat: masklint binary exits with error if files have findings
This makes it so `masklint run` displays and exits with an error when
files have findings. Previously, one would have to parse the output of
the command to see if it was empty.

Adds typing to the result to distinguish warnings (in the case of no
linter found) from actual lint failures.